feat: Implement inventory system and crafting mechanics
- Added inventory management with slots and item counts.
- Introduced crafting recipes for items including planks, workbenches, and torches.
- Updated player interactions to include mining and placing blocks through the inventory.
- Enhanced rendering to display inventory and crafting menus.
- Modified world generation to support chunk-based terrain and tile management.
- Refactored player movement and collision detection to accommodate new mechanics.
- Added new tile definitions for wood, leaves, workbench, and torch.
- Improved camera and rendering logic for better gameplay experience.

+++ include/camera.h
@@ -5,7 +5,7 @@
 #include "player.h"
 
 typedef struct Camera {
-    uint8_t x;
+    uint16_t x;
     uint8_t y;
 } Camera;
